Volume management improvements.

+++ lib/powernode/models/node_instance.rb
@@ -56,7 +56,7 @@ class NodeInstance
   end
 
   def check!
-    if instance && instance.state != 'terminated'
+    if instance
       Powernode.logger.info "Updating cloud instance #{id}."
       begin
         self.private_ip_address = instance.private_ip_address
@@ -66,8 +66,10 @@ class NodeInstance
         Powernode.logger.error "Exception: #{e.message}."
       end
       save if changed?
-    else
-      self.destroy
+      case status
+      when 'terminated'
+        self.destroy
+      end
     end
   end
 
@@ -192,11 +194,20 @@ class NodeInstance
     rescue => e
       Powernode.logger.error "Exception: #{e.message}."
     end
+    response = ''
     begin
-      session.exec!("sudo #{command}") if session
+      session.open_channel do |channel|
+        channel.exec("sudo #{command}") do |ch, success|
+          channel.on_data do |ch, data|
+            response = data
+          end
+        end
+      end
+      session.loop
     rescue => e
       Powernode.logger.error "Exception: #{e.message}."
     end
+    response
   end
 
   def netboot_sync!
